requirements.txt
django~=4.0.4
djangorestframework~=3.13.1
six~=1.16.0